+++ src/main/java/agents/firm/sales/pricing/PriceFollower.java
@@ -83,7 +83,7 @@ public class PriceFollower implements AskPricingStrategy {
      * All inventory is unwanted
      */
     @Override
-    public int estimateSupplyGap() {
+    public float estimateSupplyGap() {
         return sales.getHowManyToSell();
     }
 }
