+++ src/main/java/model/utilities/dummies/DummyBuyer.java
@@ -152,7 +152,7 @@ public class DummyBuyer extends Firm {
      * how "far" sales inventory are from target.
      */
     @Override
-    public int estimateSupplyGap(GoodType type) {
+    public float estimateSupplyGap(GoodType type) {
         return 0;
     }
 }
